refactor: decouple minute-related logic from meeting module by creating dedicated minute module and repositories
