+++ src/app.module.ts
@@ -24,6 +24,7 @@ import { ScopeGuard } from '@/auth/guards/scope.guard';
 import { PermissionGuard } from '@/admin/permission/guards/permission.guard';
 import { PrismaModule } from './prisma/prisma.module';
 import { MeetingModule } from './meeting/meeting.module';
+import { MinuteModule } from './minute/minute.module';
 
 import { TencentMtgModule } from './tencent-mtg/tencent-mtg.module';
 import { LarkMeetingModule } from './lark-meeting/lark-meeting.module';
@@ -106,6 +107,7 @@ import { TrackingReportModule } from './tracking-report/tracking-report.module';
     AuthModule,
     UserModule,
     MeetingModule,
+    MinuteModule,
 
     TencentMtgModule,
     LarkMeetingModule,
