+++ src/mcp-server/mcp-server.module.ts
@@ -13,7 +13,7 @@ import { Module } from '@nestjs/common';
 import { McpModule } from '@rekog/mcp-nest';
 import { RoleModule } from '@/admin/role/role.module';
 import { MeetingModule } from '@/meeting/meeting.module';
-import { MinuteParticipantSummaryRepository } from '@/meeting/repositories';
+import { MinuteParticipantSummaryRepository } from '@/minute/repositories';
 
 import {
   GreetingTool,
