+++ src/meeting/decorators/index.ts
@@ -1,0 +1,1 @@
+export * from './meeting.decorators';
