+++ src/meeting/dto/index.ts
@@ -1,12 +1,8 @@
 export * from './meeting-record-query.dto';
 export * from './meeting-record-response.dto';
 export * from './meeting-record-stats.dto';
-export * from './meeting-summary.dto';
-export * from './participant-summary.dto';
 export * from './meeting-participant.dto';
 export * from './meeting-record-delete.dto';
 export * from './meeting-record-create.dto';
 export * from './meeting-record-update.dto';
-export * from './transcript.dto';
 export * from './health-check.dto';
-export * from './meeting-recording.dto';
