+++ src/meeting/repositories/index.ts
@@ -1,8 +1,3 @@
 export { MeetingRepository } from './meeting.repository';
-export { MeetingFileRepository } from './meeting-file.repository';
-export { MinuteSummaryRepository } from './meeting-summary.repository';
-export { MinuteRepository } from './meeting-recording.repository';
-export { TranscriptRepository } from './transcript.repository';
 
 export { MeetingParticipantRepository } from './meeting-participant.repository';
-export { MinuteParticipantSummaryRepository } from './participant-summary.repository';
