+++ src/meeting/services/index.ts
@@ -1,5 +1,1 @@
 export * from './meeting.service';
-export * from './meeting-recording.service';
-export * from './meeting-summary.service';
-export * from './transcript.service';
-export { ParticipantSummaryService } from './participant-summary.service';
