+++ src/meeting/types/index.ts
@@ -1,5 +1,1 @@
-// 会议记录相关类型
-export * from './meeting-record.types';
-
-// 会议文件相关类型
-export * from './meeting-file.types';
+export * from './meeting.types';
