+++ src/minute/controllers/index.ts
@@ -1,0 +1,4 @@
+export * from './minute.controller';
+export * from './minute-summary.controller';
+export * from './minute-participant-summary.controller';
+export * from './transcript.controller';
