+++ src/minute/dto/index.ts
@@ -1,0 +1,4 @@
+export * from './minute.dto';
+export * from './minute-summary.dto';
+export * from './minute-participant-summary.dto';
+export * from './transcript.dto';
