+++ src/minute/dto/minute-participant-summary.dto.ts
