+++ src/minute/repositories/index.ts
@@ -1,0 +1,5 @@
+export * from './minute.repository';
+export * from './minute-summary.repository';
+export * from './minute-participant-summary.repository';
+export * from './minute-file.repository';
+export * from './transcript.repository';
