+++ src/minute/repositories/minute-participant-summary.repository.ts
