+++ src/minute/services/index.ts
@@ -1,0 +1,5 @@
+export * from './minute.service';
+export * from './minute-summary.service';
+export * from './minute-participant-summary.service';
+export * from './minute-participant-summary-crud.service';
+export * from './transcript.service';
