+++ src/minute/types/index.ts
@@ -1,0 +1,1 @@
+export * from './minute-file.types';
